Refactor remove DI and delegate, Include all samples, fix AspNet sample references
