+++ Hystrix.Dotnet.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26730.16
+VisualStudioVersion = 15.0.27004.2010
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{148A131E-9B31-42A7-9043-004954218C0A}"
 EndProject
@@ -36,7 +36,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hystrix.Dotnet.WebConfigura
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{8E7B5B99-5040-41F4-A79B-804D5954B874}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hystrix.Dotnet.Samples.Owin", "samples\Hystrix.Dotnet.Samples.Owin\Hystrix.Dotnet.Samples.Owin.csproj", "{0B27395A-8391-4077-805E-E7CBCBAE2152}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hystrix.Dotnet.Samples.Owin", "samples\Hystrix.Dotnet.Samples.Owin\Hystrix.Dotnet.Samples.Owin.csproj", "{0B27395A-8391-4077-805E-E7CBCBAE2152}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "common", "common", "{895C2BB2-C54B-44DC-82E6-A2ADDE94D6F4}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hystrix.Dotnet.Samples.AspNetCore", "samples\Hystrix.Dotnet.Samples.AspNetCore\Hystrix.Dotnet.Samples.AspNetCore.csproj", "{89E7E92C-5C0C-47D8-8D4E-12B194B52C1C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hystrix.Dotnet.Samples.AspNet", "samples\Hystrix.Dotnet.Samples.AspNet\Hystrix.Dotnet.Samples.AspNet\Hystrix.Dotnet.Samples.AspNet.csproj", "{8FA6FAC4-4B5D-4846-8D85-11FF0F1FB5B1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -84,6 +90,14 @@ Global
 		{0B27395A-8391-4077-805E-E7CBCBAE2152}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0B27395A-8391-4077-805E-E7CBCBAE2152}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0B27395A-8391-4077-805E-E7CBCBAE2152}.Release|Any CPU.Build.0 = Release|Any CPU
+		{89E7E92C-5C0C-47D8-8D4E-12B194B52C1C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{89E7E92C-5C0C-47D8-8D4E-12B194B52C1C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{89E7E92C-5C0C-47D8-8D4E-12B194B52C1C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{89E7E92C-5C0C-47D8-8D4E-12B194B52C1C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8FA6FAC4-4B5D-4846-8D85-11FF0F1FB5B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8FA6FAC4-4B5D-4846-8D85-11FF0F1FB5B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8FA6FAC4-4B5D-4846-8D85-11FF0F1FB5B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8FA6FAC4-4B5D-4846-8D85-11FF0F1FB5B1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -99,6 +113,8 @@ Global
 		{022AF60C-5950-4FE1-BF64-53F5BDF505A1} = {148A131E-9B31-42A7-9043-004954218C0A}
 		{B25E6D7C-D97A-419A-A3CF-1B49675B6BDF} = {E1CB0C26-CFBE-4D30-AD4D-C72DF5754AD4}
 		{0B27395A-8391-4077-805E-E7CBCBAE2152} = {8E7B5B99-5040-41F4-A79B-804D5954B874}
+		{89E7E92C-5C0C-47D8-8D4E-12B194B52C1C} = {8E7B5B99-5040-41F4-A79B-804D5954B874}
+		{8FA6FAC4-4B5D-4846-8D85-11FF0F1FB5B1} = {8E7B5B99-5040-41F4-A79B-804D5954B874}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {17C13872-5204-4DA2-974C-A9DA56B838A4}
