+++ samples/Hystrix.Dotnet.Samples.Owin/Hystrix.Dotnet.Samples.Owin.csproj
@@ -7,7 +7,6 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNet.WebApi.OwinSelfHost" Version="5.2.3" />
-    <PackageReference Include="SimpleInjector.Integration.WebApi" Version="4.0.12" />
   </ItemGroup>
 
   <ItemGroup>
